refactor: update global-not-found.tsx
